% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.eval_dots_arg}
\alias{.eval_dots_arg}
\title{evaluate dots argument}
\usage{
.eval_dots_arg(..., flatten = FALSE)
}
\arguments{
\item{...}{dots passed from function input}

\item{flatten}{should lists be flattened}
}
\value{
vector
}
\description{
evaluate dots argument
}
\details{
Make list from dots, unlist and evaluate.
}
