% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.i_find_in}
\alias{.i_find_in}
\title{internal replacement of match function for remove missing values (match is much slower but can handle more data classes)}
\usage{
.i_find_in(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
description description
}
