% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.init}
\alias{.init}
\title{backend for i_labelled}
\usage{
.init(
  x,
  label = NULL,
  labels = NULL,
  na_values = NULL,
  na_range = NULL,
  scale = NULL,
  annotation = NULL,
  wording = NULL,
  ...
)
}
\arguments{
\item{x}{vector}

\item{label}{variable label}

\item{labels}{value labels as named vector}

\item{na_values}{missing values (e.g. c(888, 999))}

\item{na_range}{range of missing values (e.g. c(-9,-1))}

\item{scale}{scale level (nominal, ordinal, scale)}

\item{annotation}{additional information about variable}

\item{wording}{question text}

\item{...}{further attributes passed to class}
}
\value{
x as i_labelled object with attributes applied to it
}
\description{
all arguments are passed from i_labelled
}
