% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{.valid_annotation}
\alias{.valid_annotation}
\title{validate annotation - intern}
\usage{
.valid_annotation(x)
}
\arguments{
\item{x}{character vector or NULL}
}
\value{
T/F
}
\description{
contains run-time-tests annotation
runs internally
}
