% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{.valid_labels}
\alias{.valid_labels}
\title{validate value labels - intern}
\usage{
.valid_labels(x)
}
\arguments{
\item{x}{named vector (label = value)}
}
\value{
No return value. Aborts process when run-time-tests fail.
}
\description{
contains several run-time-tests for value labels
runs internally
}
