% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{.valid_missing}
\alias{.valid_missing}
\title{validate missing values/range - intern}
\usage{
.valid_missing(x)
}
\arguments{
\item{x}{vector}
}
\value{
T/F
}
\description{
validate missing values/range - intern
}
