% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{.valid_scale}
\alias{.valid_scale}
\title{validate scale label - intern}
\usage{
.valid_scale(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Aborts process when scale level invalid.
}
\description{
run-time-tests for scale level
runs internally
}
