% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{.valid_subject}
\alias{.valid_subject}
\title{validate subject - intern}
\usage{
.valid_subject(x)
}
\arguments{
\item{x}{character vector or NULL}
}
\value{
T/F
}
\description{
contains run-time-tests subject
runs internally
}
