% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wording.R
\name{.valid_wording}
\alias{.valid_wording}
\title{validate wording - intern}
\usage{
.valid_wording(x)
}
\arguments{
\item{x}{character vector or NULL}
}
\value{
T/F
}
\description{
contains run-time-tests wording
runs internally
}
