% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{\%in\%}
\alias{\%in\%}
\title{find matches in vector (return T/F)}
\usage{
x \%in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\description{
find matches in vector (return T/F)
}
