% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{methods}
\name{grapesingrapes}
\alias{grapesingrapes}
\alias{\%in\%}
\alias{\%in\%,i_labelled,ANY-method}
\alias{\%in\%,ANY,i_labelled-method}
\alias{\%in\%,i_labelled,i_labelled-method}
\title{generic for \%in\%}
\usage{
x \%in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
T/F
}
\description{
generic for \%in\%
}
