% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{i_annotation}
\alias{i_annotation}
\title{add annotation to variable}
\usage{
i_annotation(x, annotation, overwrite = FALSE)
}
\arguments{
\item{x}{vector}

\item{annotation}{variable label as string or NULL (NULL will remove label)}

\item{overwrite}{overwrite existing annotation and replace with new annotation}
}
\value{
x with annotation applied
}
\description{
add annotation to i_labelled object

can be used to store additional information about a variable
}
