% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{i_as_character}
\alias{i_as_character}
\title{as character}
\usage{
i_as_character(
  x,
  missing_to_na = FALSE,
  require_all_labels = FALSE,
  keep_attributes = FALSE
)
}
\arguments{
\item{x}{vector}

\item{missing_to_na}{as missing declared values will become NA}

\item{require_all_labels}{process will be interrupted, when not all values have valid labels}

\item{keep_attributes}{should attributes be preserved}
}
\value{
Character vector
}
\description{
make character from i_labelled
}
