% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{i_as_factor}
\alias{i_as_factor}
\title{as factor}
\usage{
i_as_factor(x, missing_to_na = F, require_all_labels = F, keep_attributes = F)
}
\arguments{
\item{x}{vector}

\item{missing_to_na}{as missing declared values will become NA}

\item{require_all_labels}{process will be interrupted, when not all values have valid labels}

\item{keep_attributes}{should attributes be preserved}
}
\description{
make factor from i_labelled
}
