% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{i_as_numeric}
\alias{i_as_numeric}
\title{as numeric}
\usage{
i_as_numeric(x, missing_to_na = FALSE, keep_attributes = FALSE)
}
\arguments{
\item{x}{vector}

\item{missing_to_na}{as missing declared values will become NA}

\item{keep_attributes}{should attributes be preserved}
}
\value{
numeric vector
}
\description{
make numeric from i_labelled
}
