% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_assert_labels}
\alias{i_assert_labels}
\title{Check for required value labels in set of variables}
\usage{
i_assert_labels(x, labels, info = NULL, verbose = T)
}
\arguments{
\item{x}{data.frame}

\item{labels}{character vector}

\item{info}{string with info message (purpose of assertion) - optional}

\item{verbose}{return TRUE when assertion is successful}
}
\description{
Check for required value labels in set of variables
}
