% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to.R
\name{i_data_to_double}
\alias{i_data_to_double}
\title{data to double}
\usage{
i_data_to_double(x)
}
\arguments{
\item{x}{vector}
}
\description{
changes the data within class i_labelled

will also change base R classes, since it is only a wrapper for keeping attributes
}
