% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{i_data_to_integer}
\alias{i_data_to_integer}
\title{data to integer}
\usage{
i_data_to_integer(x)
}
\arguments{
\item{x}{vector}
}
\description{
changes the data within class i_labelled

will also change base R classes, since it is only a wrapper for keeping attributes
}
