% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_annotation}
\alias{i_get_annotation}
\title{get annotation}
\usage{
i_get_annotation(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
returns annotation
}
\description{
return annotation as character vector applied to vector
return list when applied to data.frame
}
