% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_equal_subject}
\alias{i_get_equal_subject}
\title{get variable names by subject}
\usage{
i_get_equal_subject(x, subject = NULL)
}
\arguments{
\item{x}{data.frame}

\item{subject}{one or more subjects as character vector. when NULL return all variable names by all subjects in data}
}
\value{
named list or NA. return named list with one list entry for each subject. when no subject in data or no match for subjects, return NA.
}
\description{
return all variable names by subjects

one or more subjects can be looked up
}
