% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_equal_wording}
\alias{i_get_equal_wording}
\title{get variable names by wording}
\usage{
i_get_equal_wording(x, wording = NULL)
}
\arguments{
\item{x}{data.frame}

\item{wording}{one or more wordings as character vector. when NULL return all variable names by all wordings in data}
}
\value{
named list or NA. return named list with one list entry for each wording. when no wording in data or no match for wordings, return NA.
}
\description{
return all variable names by wordings

one, several, or all wordings can be looked up
}
