% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_label}
\alias{i_get_label}
\title{get variable label}
\usage{
i_get_label(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Variable label
}
\description{
return variable label when applied to vector
return list when applied to data.frame
}
