% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_get_na_range}
\alias{i_get_na_range}
\title{print missing range}
\usage{
i_get_na_range(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\description{
print missing range when applied to vector
return list when applied to data.frame
}
