% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_na_values}
\alias{i_get_na_values}
\title{get missing values}
\usage{
i_get_na_values(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Return missing values
}
\description{
return missing values when applied to vector
return list when applied to data.frame
}
