% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_subject}
\alias{i_get_subject}
\title{get subject}
\usage{
i_get_subject(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
returns subject
}
\description{
return subject as character vector applied to vector
return list when applied to data.frame
}
