% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_wording}
\alias{i_get_wording}
\title{get wording}
\usage{
i_get_wording(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
returns wording
}
\description{
return wording as character vector applied to vector
return list when applied to data.frame
}
