% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{i_in}
\alias{i_in}
\title{Match values in i_labelled data via value labels}
\usage{
i_in(x, table)
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\description{
Find matches in vector (return T/F)
}
