% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{i_label}
\alias{i_label}
\title{set variable label}
\usage{
i_label(x, label)
}
\arguments{
\item{x}{vector}

\item{label}{variable label as string or NULL (NULL will remove label)}
}
\value{
x with variable label applied
}
\description{
set variable label
}
