% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{i_labelled.data.frame}
\alias{i_labelled.data.frame}
\title{make all variables in data.frame i_labelled (S3 method)}
\usage{
\method{i_labelled}{data.frame}(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
make all variables in data.frame i_labelled (S3 method)
}
