% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{i_labelled}
\alias{i_labelled}
\title{class constructor}
\usage{
i_labelled(
  x,
  label = NULL,
  labels = NULL,
  na_values = NULL,
  na_range = NULL,
  ...
)
}
\arguments{
\item{x}{vector}

\item{label}{variable label}

\item{labels}{value labels as named vector (e.g. c("A"=1, "B"=2) or setNames(c(1,2), c("A","B")))}

\item{na_values}{missing values (e.g. c(888, 999))}

\item{na_range}{range of missing values as vector length 2 (e.g. c(-9,-1))}

\item{...}{further attributes passed to class}
}
\description{
class constructor
}
