% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_labels}
\alias{i_labels}
\title{set value labels}
\usage{
i_labels(x, ..., overwrite = FALSE)
}
\arguments{
\item{x}{vector}

\item{...}{set labels for values (e.g. label_of_choice = 1 or "Label of Choice" = 1); remove single label with NULL = value (e.g. NULL = 1); removes all value labels when only NULL (e.g. i_label(x, NULL))}

\item{overwrite}{when TRUE, all existing labels are dropped and replaced with the new ones}
}
\value{
returns x with value labels applied
}
\description{
set value labels
}
\details{
In order to assign a specific label to multiple values a named list can also be provided to ... (e.g. list(missing = -9:-1, valid = 1:3))

A named vector can also be provided (e.g. setNames(c(1,2), c("A","B")))
}
