% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{i_missing_to_na}
\alias{i_missing_to_na}
\title{missing values to NA}
\usage{
i_missing_to_na(x, remove_missing_labels = F)
}
\arguments{
\item{x}{vector or data.frame}

\item{remove_missing_labels}{remove values labels from values which are declared as missing}
}
\description{
alle values declared as missing will be recoded as NA
set missing values via i_labelled(), i_na_values() or i_na_range()
}
