% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{i_na_values}
\alias{i_na_values}
\title{define missing values}
\usage{
i_na_values(x, values, sort = T, desc = F)
}
\arguments{
\item{x}{vector}

\item{values}{vector with missing values e.g. c(888,999) or NULL (NULL will remove all missing values)}

\item{sort}{sort values}

\item{desc}{sort values in descending order}
}
\description{
define which values will be handled as missing values
}
