% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_annotation}
\alias{i_print_annotation}
\title{print annotation}
\usage{
i_print_annotation(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print annotation attribute to console
}
\description{
print annotation
}
