% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_attributes}
\alias{i_print_attributes}
\title{print attributes}
\usage{
i_print_attributes(x, exclude = NULL)
}
\arguments{
\item{x}{vector}

\item{exclude}{character vector with attribute names not taken into account}
}
\value{
No return value. Print attributes to console
}
\description{
print attributes
}
