% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_label}
\alias{i_print_label}
\title{print variable label}
\usage{
i_print_label(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print variable label to console
}
\description{
print variable label
}
