% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_labels}
\alias{i_print_labels}
\title{print value labels}
\usage{
i_print_labels(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print labels to console
}
\description{
print value labels
}
