% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_na_values}
\alias{i_print_na_values}
\title{print missing values}
\usage{
i_print_na_values(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print na values to console
}
\description{
print missing values
}
