% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_scale}
\alias{i_print_scale}
\title{print scale level}
\usage{
i_print_scale(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print scale level to console
}
\description{
print scale level
}
