% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_subject}
\alias{i_print_subject}
\title{print subject}
\usage{
i_print_subject(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print subject attribute to console
}
\description{
print subject
}
