% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{i_print_wording}
\alias{i_print_wording}
\title{print wording}
\usage{
i_print_wording(x)
}
\arguments{
\item{x}{vector}
}
\value{
No return value. Print wording attribute to console
}
\description{
print wording
}
