% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{i_recode}
\alias{i_recode}
\title{Function for recoding new variable from origin vairiable.
Does not copy old variable, but creates new variable.}
\usage{
i_recode(x, ..., label = NULL, na_values = NULL, na_range = NULL)
}
\arguments{
\item{x}{vector}

\item{...}{formular for recoding of values as follows: newValueLabel = newValue ~ c(oldValues)}

\item{label}{variable label}

\item{na_values}{a vector with missing values}

\item{na_rage}{a vector for missing range}
}
\description{
Function for recoding new variable from origin vairiable.
Does not copy old variable, but creates new variable.
}
\examples{
i_recode(i_labelled(1:4, labels = c("A" = 1, "B" = 2, "C" = 3, "D" = 4)), "AB" = 1 ~ c("A", "B"), "CD" = 2 ~ c(3, 4))

}
