% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_annotation}
\alias{i_remove_annotation}
\title{remove annotation}
\usage{
i_remove_annotation(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without annotation
}
\description{
remove annotation label from variable
keep other attributes
}
