% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_label}
\alias{i_remove_label}
\title{remove variable label}
\usage{
i_remove_label(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without variable label
}
\description{
remove variable label
keep other attributes
}
