% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_labels}
\alias{i_remove_labels}
\title{remove all value labels}
\usage{
i_remove_labels(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without value labels
}
\description{
remove all value labels
keep other attributes
}
