% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_na_values}
\alias{i_remove_na_values}
\title{remove as na values}
\usage{
i_remove_na_values(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without na-values
}
\description{
remove na values (information which values should be handled as missing)
keep other attributes
}
