% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_scale}
\alias{i_remove_scale}
\title{remove scale level}
\usage{
i_remove_scale(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without scale level
}
\description{
remove scale label from variable
keep other attributes
}
