% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_subject}
\alias{i_remove_subject}
\title{remove subject}
\usage{
i_remove_subject(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without subject
}
\description{
remove subject label from variable
keep other attributes
}
