% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_wording}
\alias{i_remove_wording}
\title{remove wording}
\usage{
i_remove_wording(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without wording
}
\description{
remove wording label from variable
keep other attributes
}
