% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{i_scale}
\alias{i_scale}
\title{set scale level}
\usage{
i_scale(x, scale)
}
\arguments{
\item{x}{vector}

\item{scale}{scale level (nominal, ordinal, scale) as string or NULL (NULL will remove scale level)}
}
\value{
Returns x with scale label set
}
\description{
set scale level
}
