% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_sort_labels}
\alias{i_sort_labels}
\title{sort value labels by values or by labels}
\usage{
i_sort_labels(x, by = "values", decreasing = F)
}
\arguments{
\item{x}{vector or data.frame}

\item{by}{either values or labels}

\item{decreasing}{sort decreasing}
}
\description{
sort value labels by values or by labels
}
