% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{i_subject}
\alias{i_subject}
\title{add subject to variable}
\usage{
i_subject(x, subject)
}
\arguments{
\item{x}{vector}

\item{subject}{variable label as string or NULL (NULL will remove label)}
}
\value{
x with subject applied
}
\description{
add subject to i_labelled object
}
