% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{i_table}
\alias{i_table}
\title{cross tabulation and table creation using i_labelled labels}
\usage{
i_table(..., missing_to_na = TRUE, as_factor = TRUE, table_args = NULL)
}
\arguments{
\item{...}{one or more atomic vectors}

\item{missing_to_na}{make as missing declared values NA}

\item{as_factor}{make labelled data factor before pass to table}

\item{table_args}{arguments of base::table as named list}
}
\value{
returns a contingency table, an object of class "table"
}
\description{
wrapper for base::table

convert i_labelled objects to base class and pass to table function
}
