% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{i_valid_annotation}
\alias{i_valid_annotation}
\title{validate annotation}
\usage{
i_valid_annotation(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
T/F
}
\description{
returns boolean when applied to vector

returns a named list when applied to data.frame
}
