% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_valid_labels}
\alias{i_valid_labels}
\title{validate value labels}
\usage{
i_valid_labels(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
No return value. Aborts process when run-time-tests fail
}
\description{
returns boolean when applied to vector

returns a named list when applied to data.frame
}
