% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{i_valid_scale}
\alias{i_valid_scale}
\title{validate variable scale level}
\usage{
i_valid_scale(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
T/F
}
\description{
returns boolean when applied to vector

returns a named list when applied to data.frame
}
