% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{i_valid_subject}
\alias{i_valid_subject}
\title{validate subject}
\usage{
i_valid_subject(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
T/F
}
\description{
returns boolean when applied to vector

returns a named list when applied to data.frame
}
