% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wording.R
\name{i_valid_wording}
\alias{i_valid_wording}
\title{validate wording}
\usage{
i_valid_wording(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
T/F
}
\description{
returns boolean when applied to vector

returns a named list when applied to data.frame
}
