% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wording.R
\name{i_wording}
\alias{i_wording}
\title{add wording to variable}
\usage{
i_wording(x, wording)
}
\arguments{
\item{x}{vector}

\item{wording}{variable label as string or NULL (NULL will remove label)}
}
\value{
x with wording applied
}
\description{
add wording to i_labelled object

can be used to store additional information about a variable
}
