% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{is.i_labelled}
\alias{is.i_labelled}
\title{make all variables in data.frame i_labelled (dedicated function)}
\usage{
is.i_labelled(x)
}
\arguments{
\item{x}{vector of class i_labelled}
}
\value{
x with all variables coerced to i_labelled object

T/F
}
\description{
make all variables in data.frame i_labelled (dedicated function)
}
