% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_decimal}
\alias{is_decimal}
\title{Check if vector contains decimal values}
\usage{
is_decimal(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
T/F
}
\description{
Check if vector contains decimal values
}
