% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Date}
\alias{print.Date}
\title{custom print method for Date}
\usage{
\method{print}{Date}(x, ...)
}
\arguments{
\item{x}{vector of class Date}

\item{...}{not used}
}
\description{
custom print method for Date
}
