% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.i_labelled}
\alias{print.i_labelled}
\title{custom print method for i_labelled}
\usage{
\method{print}{i_labelled}(x, ...)
}
\arguments{
\item{x}{vector of class i_labelled}

\item{...}{not used}
}
\value{
No return value. Print object data and information to console
}
\description{
custom print method for i_labelled
}
