% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{[.i_labelled}
\alias{[.i_labelled}
\title{subsetting vectors of class i_labelled}
\usage{
\method{[}{i_labelled}(x, ...)
}
\arguments{
\item{x}{vector of class i_labelled}

\item{...}{not used}
}
\description{
subsetting vectors of class i_labelled
}
