% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.i_labelled}
\alias{as.i_labelled}
\title{as.i_labelled}
\usage{
as.i_labelled(x, ...)
}
\arguments{
\item{x}{vector or data.frame}

\item{...}{attributes passed to class}
}
\description{
coerce to i_labelled class

applicable to vector or data.frame objects
}
