% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.i_labelled}
\alias{as.i_labelled}
\title{coerce to i_labelled class}
\usage{
as.i_labelled(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{attributes passed to class}
}
\value{
vector of class i_labelled
}
\description{
coerce to i_labelled class
}
