% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{.merge_labels}
\alias{.merge_labels}
\title{combine old value labels with new value labels}
\usage{
.merge_labels(old_labs, new_labs, sort_desc = F)
}
\arguments{
\item{old_labs}{named vector}

\item{new_labs}{named vector}

\item{sort_desc}{sort value labels in descending order according to values}
}
\description{
return named vector
}
