% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{\%in\%}
\alias{\%in\%}
\title{Match values in i_labelled data via value labels}
\usage{
x \%in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
Vector of T/F values with length of x
}
\description{
Find matches in vector (return T/F)
}
