% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to.R
\name{i_as_factor}
\alias{i_as_factor}
\title{as factor}
\usage{
i_as_factor(
  x,
  labels = NULL,
  require_all_labels = T,
  missing_to_na = F,
  remove_missing_labels = F,
  ordered = F
)
}
\arguments{
\item{x}{vector}

\item{labels}{value labels as named vector (e.g. c("A"=1, "B"=2) or setNames(c(1,2), c("A","B")))}

\item{require_all_labels}{process will be interrupted, when not all values have valid labels}

\item{missing_to_na}{as missing declared values will become NA}

\item{remove_missing_labels}{as missing declared values will be removed from levels}

\item{ordered}{order labels by the underlying values}
}
\description{
make factor from i_labelled
}
