% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{i_as_factor}
\alias{i_as_factor}
\title{as factor}
\usage{
i_as_factor(x, labels = NULL)
}
\arguments{
\item{x}{vector}

\item{labels}{value labels as named vector or named list (e.g. list("A"=1, "B"=2) or c("A"=1, "B"=2) or setNames(c(1,2), c("A","B")))}
}
\description{
make factor from i_labelled
}
