% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_assert_labels}
\alias{i_assert_labels}
\title{Check for required value labels in set of variables}
\usage{
i_assert_labels(x, variables, labels, info = NULL, verbose = T)
}
\arguments{
\item{x}{data.frame}

\item{variables}{character vector}

\item{labels}{character vector}
}
\description{
Check for required value labels in set of variables
}
