% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{i_copy}
\alias{i_copy}
\title{copy labels from one variable to another}
\usage{
i_copy(to, from, what = "all", ...)
}
\arguments{
\item{to}{vector}

\item{from}{vector}

\item{what}{character vector describing which labels are copied: 'all' (default), 'label', 'labels', 'na_values', 'na_range'}

\item{...}{further attributes passed to structure}
}
\description{
copy labels from one variable to another
}
