% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to.R
\name{i_data_to_numeric}
\alias{i_data_to_numeric}
\title{data to numeric}
\usage{
i_data_to_numeric(x)
}
\arguments{
\item{x}{vector}
}
\value{
Returns x coerced to numeric
}
\description{
changes the data within class i_labelled

will also change base R classes, since it is only a wrapper for keeping attributes
}
