% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{i_get_scale}
\alias{i_get_scale}
\title{get scale level}
\usage{
i_get_scale(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns scale level
}
\description{
return scale level when applied to vector
return list when applied to data.frame
}
