% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{i_labelled_df}
\alias{i_labelled_df}
\title{make all variables in data.frame i_labelled}
\usage{
i_labelled_df(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
make all variables in data.frame i_labelled
}
