% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{i_labels}
\alias{i_labels}
\title{set value labels}
\usage{
i_labels(x, ..., overwrite = FALSE)
}
\arguments{
\item{x}{vector}

\item{...}{set labels for values (e.g. label_of_choice = 1 or "Label of Choice" = 1); remove single label with NULL = value (e.g. NULL = 1); removes all value labels when only NULL (e.g. i_label(x, NULL))}

\item{overwrite}{should new labels be merged with existing labels or remove existing labels}
}
\description{
set value labels
}
