% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{i_recode}
\alias{i_recode}
\title{Function for recoding new variable from origin vairiable.
Does not copy old variablen, but creates new variable.}
\usage{
i_recode(x, ..., label = NULL, na_values = NULL, na_range = NULL)
}
\arguments{
\item{x}{vector}

\item{...}{formular for recoding of values}

\item{label}{variable label}

\item{na_values}{a vector with missing values}

\item{na_rage}{a vector for missing range}
}
\description{
Function for recoding new variable from origin vairiable.
Does not copy old variablen, but creates new variable.
}
\examples{
# i_recode(myData$oldVar, "newLab1" = 1 = c("oldLab1", "oldLab2"))

}
