% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{i_remove_na_range}
\alias{i_remove_na_range}
\title{remove as na range}
\usage{
i_remove_na_range(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
Returns x without na-range
}
\description{
remove na range (information which values should be handled as missing)
keep other attributes
}
