% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_base_r.R
\name{i_to_base_class}
\alias{i_to_base_class}
\title{remove class i_labelled and return base R class}
\usage{
i_to_base_class(x, as_factor = T, keep_attributes = F)
}
\arguments{
\item{x}{vector or data.frame}

\item{as_factor}{convert to factor, when all value labels are available}

\item{keep_attributes}{should attributes be preserved}
}
\description{
- when value labels for all values are available will return factor
- when value labels are missing will unclass i_labelled
- remove class i_labelled and return variable as base R class
}
