% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{i_to_base_class}
\alias{i_to_base_class}
\title{remove class i_labelled and return base R class}
\usage{
i_to_base_class(
  x,
  missing_to_na = TRUE,
  as_factor = TRUE,
  keep_attributes = FALSE
)
}
\arguments{
\item{x}{vector or data.frame}

\item{missing_to_na}{missing values will become regular NA}

\item{as_factor}{convert to factor when value labels are available}

\item{keep_attributes}{should attributes be preserved}
}
\value{
Returns x coerced to R base class
}
\description{
- when value labels for all values are available will return factor
- when value labels are missing will unclass i_labelled
- remove class i_labelled and return variable as base R class
}
