% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{i_unclass}
\alias{i_unclass}
\title{unclass variables}
\usage{
i_unclass(x, keep_attributes = FALSE)
}
\arguments{
\item{x}{vector or data.frame}

\item{keep_attributes}{should attributes be preserved}
}
\value{
x unclassed
}
\description{
unclass variables
}
