% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{is.i_labelled}
\alias{is.i_labelled}
\title{check for class i_labelled}
\usage{
is.i_labelled(x)
}
\arguments{
\item{x}{vector of class i_labelled}
}
\value{
T/F
}
\description{
check for class i_labelled
}
