% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.i_labelled}
\alias{print.i_labelled}
\title{custom print method}
\usage{
\method{print}{i_labelled}(x, ...)
}
\arguments{
\item{x}{array of class i_labelled}

\item{...}{not used}
}
\description{
custom print method
}
